+++ src/template/index.ts
@@ -1,13 +1,17 @@
 import { TemplatesRepository } from './infra/repositories/TemplatesRepository'
 import { CreateDatasetTemplate } from './domain/useCases/CreateDatasetTemplate'
+import { DeleteTemplate } from './domain/useCases/DeleteTemplate'
 import { GetDatasetTemplates } from './domain/useCases/GetDatasetTemplates'
+import { GetTemplate } from './domain/useCases/GetTemplate'
 
 const templatesRepository = new TemplatesRepository()
 
 const createDatasetTemplate = new CreateDatasetTemplate(templatesRepository)
+const deleteTemplate = new DeleteTemplate(templatesRepository)
 const getDatasetTemplates = new GetDatasetTemplates(templatesRepository)
+const getTemplate = new GetTemplate(templatesRepository)
 
-export { createDatasetTemplate, getDatasetTemplates }
+export { createDatasetTemplate, deleteTemplate, getDatasetTemplates, getTemplate }
 export {
   CreateDatasetTemplateDTO,
   TemplateFieldDTO,