+++ src/template/infra/repositories/TemplatesRepository.ts
@@ -4,7 +4,10 @@ import { CreateDatasetTemplateDTO } from '../../domain/dtos/CreateDatasetTemplat
 import { DatasetTemplate } from '../../domain/models/DatasetTemplate'
 import { ITemplatesRepository } from '../../domain/repositories/ITemplatesRepository'
 import { DatasetTemplatePayload } from './transformers/DatasetTemplatePayload'
-import { transformDatasetTemplatePayloadToDatasetTemplate } from './transformers/datasetTemplateTransformers'
+import {
+  transformTemplatePayloadToTemplate,
+  transformTemplatePayloadsToTemplates
+} from './transformers/datasetTemplateTransformers'
 
 export class TemplatesRepository extends ApiRepository implements ITemplatesRepository {
   private readonly collectionsResourceName: string = 'dataverses'
@@ -23,13 +26,31 @@ export class TemplatesRepository extends ApiRepository implements ITemplatesRepo
       })
   }
 
+  public async getTemplate(templateId: number): Promise<DatasetTemplate> {
+    return this.doGet(`/dataverses/${templateId}/template`, true)
+      .then((response: AxiosResponse<{ data: DatasetTemplatePayload }>) =>
+        transformTemplatePayloadToTemplate(response.data.data)
+      )
+      .catch((error) => {
+        throw error
+      })
+  }
+
   public async getDatasetTemplates(
     collectionIdOrAlias: number | string
   ): Promise<DatasetTemplate[]> {
     return this.doGet(`/${this.collectionsResourceName}/${collectionIdOrAlias}/templates`, true)
       .then((response: AxiosResponse<{ data: DatasetTemplatePayload[] }>) =>
-        transformDatasetTemplatePayloadToDatasetTemplate(response.data.data)
+        transformTemplatePayloadsToTemplates(response.data.data)
       )
+      .catch((error) => {
+        throw error
+      })
+  }
+
+  public async deleteTemplate(templateId: number): Promise<void> {
+    return this.doDelete(`/dataverses/${templateId}/template`)
+      .then(() => undefined)
       .catch((error) => {
         throw error
       })